==> spaCy_OSM_GoogleMaps/1. Join datasets.ipynb <==
import pandas as pd
import base64
import numpy as np

import urllib
import nltk
import sys, os
from typing import List
import base64
from sqlalchemy import create_engine
import iso639
import psycopg2 as pg
pd.set_option('display.max_colwidth', 1)

# Connecting to the db

def getCursor():
    database_settings = "dbname=" + dbname + " user=" + user + " host=" + host + " port=" + port + " password=" + password

    engine_url = 'postgresql://' + user + ':' + password + '@' + host + ':'+ port +'/'+dbname
    engine = create_engine(engine_url)

    conn = pg.connect(database_settings)
    cur = conn.cursor()
    return cur,conn,engine

# Encrypting postgres password

#password = 'insert password'.encode("utf-8")
#encoded = base64.b64encode(password)

#print(encoded)

dbname = 'socialmedia'
user= 'helen'
host= 'humandata.zgis.at'
port= '5433'
file = open('humandata_key.key', 'rb')
key = file.read() # The key will be type bytes
file.close()
password = base64.b64decode(key).decode("utf-8")
cur,conn,engine = getCursor()

schema = 'movement_detection'
table = 'data_2019_final_selection'
schematable = schema + '.' + table

#schematable = "public.test_sample_location_extraction"
schematable = "movement_detection.data_2019_final_selection"
SQL = "SELECT * FROM " + schematable  
print(SQL)
data_plus = pd.read_sql_query(SQL,con=conn)

data_plus.head(2)

data_plus['source'].unique()

data_plus = data_plus.replace({'<a href="http://tapbots.com/tweetbot" rel="nofollow">Tweetbot for iΟS</a>':'ios',
                              '<a href="http://twitter.com/download/iphone" rel="nofollow">Twitter for iPhone</a>':'iphone',
                              '<a href="http://twitter.com/download/android" rel="nofollow">Twitter for Android</a>':'Android',
                              '<a href="http://instagram.com" rel="nofollow">Instagram</a>':'Instagram',
                              '<a href="http://foursquare.com" rel="nofollow">Foursquare</a>':'Foursquare',
                              '<a href="https://www.hootsuite.com" rel="nofollow">Hootsuite Inc.</a>':'Hootsuite',
                              '<a href="https://www.swarmapp.com" rel="nofollow">Foursquare Swarm</a>':'fswarm',
                              '<a href="https://untappd.com" rel="nofollow">Untappd</a>':'Untapped' })

data_plus['source'].unique()

data_plus.shape

data_plus.columns

data_plus = data_plus[['message_id', 'date', 'text', 'source','longitude',
       'latittude']]

data_plus.head(3)

data_plus.to_csv('data2/data2_complete.csv')

# Join datasets

Joining the datasets extracted from multiple sources

df = pd.read_csv('data//main3.csv')
df1 = pd.read_csv('data//fours.csv')
df2 = pd.read_csv('data//fswarm.csv')
df3 = pd.read_csv('data//insta.csv')
df4 = pd.read_csv('data//hootsuite.csv')
df5 = pd.read_csv('data//untapped.csv')

print(df.shape, df1.shape, df2.shape, df3.shape, df4.shape, df5.shape)

frames = [df, df1, df2, df3, df4, df5]

df_final = pd.concat(frames)
df_final[['text']]

df_final.columns

df_final =df_final.drop(columns = ['message_id', 'date', 'source', 'geom', 'user_id',
       'user_name'])

df_final.columns

df_final.to_csv('data//data.csv')

# Join location datasets

df = pd.read_csv('data//data_english_locations.csv')
df2 = pd.read_csv('data//data_non_english_locations1.csv')
df3 = pd.read_csv('data//data_non_english_locations2.csv')
df4 = pd.read_csv('data//data_non_english_locations3.csv')
print (df.shape, df2.shape, df3.shape, df4.shape)

df.columns

df = df.rename(columns ={'text':'en_text'})
df2 = df2.rename(columns ={'de_text':'en_text'})
df3 =df3.rename(columns ={'de_text':'en_text'})
df4 =df4.rename(columns ={'de_text':'en_text'})
df.tail()

df4.columns

df = df.drop(columns =['Unnamed: 0'])
df2 = df2.drop(columns =['Unnamed: 0','text'])
df3 = df3.drop(columns =['Unnamed: 0','text'])
df4 = df4.drop(columns =['Unnamed: 0'])
df3.head()

frames = [df2,df3,df4]

df_final = pd.concat(frames)
df_final.shape

df_final = df_final.drop_duplicates(subset =['en_text','id','latittude'], keep = 'last')
df_final.shape

df_final.to_csv('data//complete_4_locations_transla.csv')

df.head()

df_temp1 = pd.read_csv('data//locations_temp_information_full.csv')
df_temp = pd.read_csv('data//locations_temp_information_full_trans.csv')


df_temp1.columns

df_temp.columns

df_temp1.shape

df_temp.shape

frame = [df_temp1, df_temp]

temp = pd.concat(frame)

temp.shape

temp.to_csv('data//analysis_data//full_entity_extraction.csv')
