==> spaCy_OSM_GoogleMaps/2.Text cleaning.ipynb <==
# Data preprocessing

import pandas as pd

import warnings
warnings.filterwarnings('ignore')

# from util import getCursor
from IPython.display import display, HTML # import display from Ipython.display instead of IPython.core.display
import base64

# Define display limits
pd.options.display.max_colwidth = 400
pd.set_option("display.max_rows", None,)

# %matplotlib inline  
%load_ext autoreload
%autoreload 2
display(HTML("<style>.container { width:98% !important; }</style>"))

# Loading data
df = pd.read_csv('data2//data2_complete.csv')
df.shape

#Drop duplicates incase the same tweet appears twice
df.drop_duplicates(inplace=True, subset="text", keep='first')
df.shape

#Drop urls
df['text'] = df['text'].replace(r'http\S+', '', regex=True).replace(r'www\S+', '', regex=True)
df.tail()

#Drop hashtag words
#df['text'] = df['text'].replace(r'#\S+', '', regex=True)
#df.tail()

# Drop emojis
df = df.astype(str).apply(lambda x: x.str.encode('ascii', 'ignore').str.decode('ascii'))
df.tail()

# Remove multiple whitespaces, tabs and new lines
df['text'] = df['text'].str.replace('\n', '')
df['text'] = df['text'].str.replace('\t', ' ')
df['text'] = df['text'].str.replace(' {2,}', ' ', regex=True)
df['text'] = df['text'].str.strip()
df.head()

#Extract posts with 3 or more words
df['split_text']=df['text'].str.split()
df['WordsCount']=df['split_text'].apply(lambda x: len(x))
df=df[(df['WordsCount']>= 3)]
df.shape

#drop duplicates
df.drop_duplicates(inplace=True, subset="text", keep='first')
df.shape

df.to_csv('data2//data2_cleaned.csv')
