==> spaCy_OSM_GoogleMaps/3. language detect and translate.ipynb <==
# Language detect, split and Translate

# Import libriaries 
import pandas as pd

import warnings
warnings.filterwarnings('ignore')

# from util import getCursor
from IPython.display import display, HTML # import display from Ipython.display instead of IPython.core.display
import base64

# Define display limits
pd.options.display.max_colwidth = 400
pd.set_option("display.max_rows", None,)

# %matplotlib inline  
%load_ext autoreload
%autoreload 2
display(HTML("<style>.container { width:98% !important; }</style>"))

## 1. Language detect

from langdetect import detect
import re

df = pd.read_csv('data2//data2_cleaned.csv')
print (df.shape)

df = df_not_en2

# Detecting the source language and catching error for undetected cells
def detect_lang(text):
    try:
        lang = detect(text)
    except:
        lang = 'Other'
    return lang

df['lang'] = df['text'].apply(detect_lang)
df.tail(10)

## 2. Split languages 

df_en = df.loc[df['lang'] == 'en']
df_en.shape

df_en.to_csv('data2//data_english2.csv')

df_not_en = df.loc[df['lang'] != 'en']
df_not_en['lang'].nunique()

df_not_en.shape

df_not_en2 = df_not_en.loc[df_not_en['lang'] != 'Other']
df_not_en2.shape

df_not_en2.head(3)

#Drop hashtag words
df_not_en2['text'] = df_not_en2['text'].replace(r'#\S+', '', regex=True)
df_not_en2.shape

#Drop @ words
df_not_en2['text'] = df_not_en2['text'].replace(r'@\S+', '', regex=True)
df_not_en2.shape

#Drop hashtag words
df_not_en2.head(10)

#Extract posts with 3 or more words
df_not_en2['split_text']=df_not_en2['text'].str.split()
df_not_en2['WordsCount']=df_not_en2['split_text'].apply(lambda x: len(x))
df_not_en2=df_not_en2[(df_not_en2['WordsCount']>= 3)]
df_not_en2.shape

df_not_en2.head(10)

df_not_en2.to_csv('data//data_not_english.csv')

## 3. Translate 

df = pd.read_csv('data//data_not_english.csv')
df.shape

from google_trans_new import google_translator  
translator = google_translator() 

### Incase of error: *JSONDecodeError Extra Data* 
- open the google_trans_new.py installed in your environment. 
- Change line 151 in google_trans_new/google_trans_new.py which is: response = (decoded_line + ']') to response = decoded_line
- Recurrent errors; follow https://stackoverflow.com/questions/68214591/python-google-trans-new-translate-raises-error-jsondecodeerror-extra-data 

# Checking if translator works 
translate_text = translator.translate("Der Himmel ist blau und ich mag Bananen", lang_tgt='en')  
print(translate_text)

# Creating list and Index to translate and save each translated result to avoid restarting
# Make sure to run this cell only once to avoid overwritting existing results. 
translated_values = []
currentindex = 0

currentindex

for index, row in df.iloc[currentindex:].iterrows():
        translated_values.append(translator.translate(row['text'], lang_tgt = 'en'))
        print("Added  " + row['text'] )
        print("Index  " + str(currentindex) )
        currentindex = currentindex + 1
#return translated_values

df = df.drop()

df = df.head(4996)

df.shape

# Appending translated values to dataframe
df_trans = df.assign(de_text=translated_values)
df_trans.shape

df_trans[['text','de_text']].tail()

df.columns

df_trans.to_csv('data//data_non_eng_trans4996K.csv')
