==> spaCy_OSM_GoogleMaps/4. Location Extraction.ipynb <==
# Location extraction with spaCy

import numpy as np
import sys, os, pandas as pd

# Import libriaries for database connection
import base64
from sqlalchemy import create_engine
import iso639
import psycopg2 as pg

# Import libriary for natural language processing 
import nltk

# Import libriary to open URL
import urllib

# Maximise column width
pd.set_option('display.max_colwidth', None)

import warnings
warnings.filterwarnings("ignore")

# Loading the spaCy high accuracy english pretrained model
import spacy
from spacy import displacy # Displacy is used to visualise spaCy tokens
nlp =spacy.load('en_core_web_trf') # model trf higher accuracy, bigger model, slower in exercution

# Defining separate functions to extract each locational entity separately. 
def filter_location_entities(entities):
    locations = []
    for entity in entities:
        if entity.label_ == 'GPE':
                locations.append(entity)
                
    return locations

def filter_location_entities1(entities):
    locations1 = []
    for entity in entities:
        if entity.label_ == 'FAC':
                locations1.append(entity)
                      
    return locations1

def filter_location_entities2(entities):
    locations2 = []
    for entity in entities:
        if entity.label_ == 'ORG':
                locations2.append(entity)
                      
    return locations2

def filter_location_entities3(entities):
    locations3 = []
    for entity in entities:
        if entity.label_ == 'LOC':
                locations3.append(entity)
                      
    return locations3

# Load data 
df = pd.read_csv('data2//data_english.csv')
df.head(1)

df.columns

df = df.drop(columns = ['Unnamed: 0', 'Unnamed: 0.1', 'Unnamed: 0.1.1','split_text', 'WordsCount'])
df.head(2)

#Extracting locational entities
df['GPE'] = df['text'].astype(str).apply(lambda x: filter_location_entities(nlp(x).ents))
df['FAC'] = df['text'].astype(str).apply(lambda x: filter_location_entities1(nlp(x).ents))
df['ORG'] = df['text'].astype(str).apply(lambda x: filter_location_entities2(nlp(x).ents))
df['LOC'] = df['text'].astype(str).apply(lambda x: filter_location_entities3(nlp(x).ents))
df.head()

df.to_csv('data2//data2_english_locations.csv')
