==> spaCy_OSM_GoogleMaps/5. Cleaning extracted entities.ipynb <==
# Cleaning extracted entities

# Import libriaries 
import pandas as pd
import numpy as np

import warnings
warnings.filterwarnings('ignore')

# from util import getCursor
from IPython.display import display, HTML # import display from Ipython.display instead of IPython.core.display
import base64

# Define display limits
pd.options.display.max_colwidth = 400
pd.set_option("display.max_rows", None,)

# %matplotlib inline  
%load_ext autoreload
%autoreload 2
display(HTML("<style>.container { width:98% !important; }</style>"))

df = pd.read_csv('data//analysis_data//keyword+temp_filtered_posts.csv')
df.shape

df.columns

df =df.drop(columns =['Unnamed: 0', 'Unnamed: 0.1', 'Unnamed: 0.1.1.1','Unnamed: 4'])

df.head(2)

# Count distinct by source
df.groupby(['source']).size().sort_values(ascending=False)

# replace all NaN with empty string
df = df.replace(np.nan, '', regex=True)
df.head()

# Data cleaning: Remove square brakets from entities
df['GPE'] =  df['GPE'].apply(lambda x: x.replace('[','').replace(']',''))
df['FAC'] =  df['FAC'].apply(lambda x: x.replace('[','').replace(']',''))
df['ORG'] =  df['ORG'].apply(lambda x: x.replace('[','').replace(']',''))
df['LOC'] =  df['LOC'].apply(lambda x: x.replace('[','').replace(']',''))
df['TE'] =  df['TE'].apply(lambda x: x.replace('[','').replace(']',''))
df.head(2)

# Extracting rows with TE
df_TE = df[(df['TE'] != '')]
df_TE.shape

# Count distinct by source
df_TE.groupby(['source']).size().sort_values(ascending=False)

df_TE[['en_text','TE']].head(50)

# Droping rows without location entity extracted
# Only drops rows where all conditions are met that is all four entities are equal to zero

index_names = df[(df['GPE']== '') & (df['FAC']== '') & (df['ORG']== '') & (df['LOC']== '')].index
df.drop(index_names, inplace = True)
df.shape

# Count distinct by source for only posts with location extracted
df.groupby(['source']).size().sort_values(ascending=False)

# Extracting rows with TE after dropping rows without location
df_TE = df[(df['TE'] != '')]
df_TE.shape

#df.to_csv('data//analysis_data//all_locations_extracted.csv')

## Clean location entities

# Load dataset after keyword filtering and Temporal Information matching
df = pd.read_csv('data//analysis_data//keyword+temp_filtered_posts_final.csv')
df.shape

# replace all NaN with empty string
df = df.replace(np.nan, '', regex=True)
df.head()

print("Number of unique values:")
print("\n GPE =", df['GPE'].nunique(), "\n ORG =", df['ORG'].nunique(), "\n LOC =", df['LOC'].nunique(), "\n FAC =", df['FAC'].nunique())

# Remove hashtag signs on extracted locations
df['GPE'] = df['GPE'].str.replace('#','')
df['LOC'] = df['LOC'].str.replace('#','')
df['FAC'] = df['FAC'].str.replace('#','')
df['ORG'] = df['ORG'].str.replace('#','')

# Remove @ signs from extracted locations
df['GPE'] = df['GPE'].str.replace('@','')
df['LOC'] = df['LOC'].str.replace('@','')
df['FAC'] = df['FAC'].str.replace('@','')
df['ORG'] = df['ORG'].str.replace('@','')

# Add space when case changes 
df['GPE'] = df['GPE'].str.replace(r'([a-z](?=[A-Z])|[A-Z](?=[A-Z][a-z]))', r'\1 ').str.strip()
df['LOC'] = df['LOC'].str.replace(r'([a-z](?=[A-Z])|[A-Z](?=[A-Z][a-z]))', r'\1 ').str.strip()
df['FAC'] = df['FAC'].str.replace(r'([a-z](?=[A-Z])|[A-Z](?=[A-Z][a-z]))', r'\1 ').str.strip()
df['ORG'] = df['ORG'].str.replace(r'([a-z](?=[A-Z])|[A-Z](?=[A-Z][a-z]))', r'\1 ').str.strip()
df['LOC'].unique()

df[['GPE','FAC','LOC','ORG']].head(30)

# Counting comma separated entities
df['#LOC'] = df['LOC'].str.count(',')
df['#GPE'] = df['GPE'].str.count(',')
df['#ORG'] = df['ORG'].str.count(',')
df['#FAC'] = df['FAC'].str.count(',')
df[['GPE','#GPE']].head(10)

# Remove spaces between entities to generate unique entities
df['GPE'] = df['GPE'].str.replace(" ", "")
df['FAC'] = df['FAC'].str.replace(" ", "")
df['LOC'] = df['LOC'].str.replace(" ", "")
df['ORG'] = df['ORG'].str.replace(" ", "")

# Add space when case changes 
df['GPE'] = df['GPE'].str.replace(r'([a-z](?=[A-Z])|[A-Z](?=[A-Z][a-z]))', r'\1 ').str.strip()
df['LOC'] = df['LOC'].str.replace(r'([a-z](?=[A-Z])|[A-Z](?=[A-Z][a-z]))', r'\1 ').str.strip()
df['FAC'] = df['FAC'].str.replace(r'([a-z](?=[A-Z])|[A-Z](?=[A-Z][a-z]))', r'\1 ').str.strip()
df['ORG'] = df['ORG'].str.replace(r'([a-z](?=[A-Z])|[A-Z](?=[A-Z][a-z]))', r'\1 ').str.strip()

#Keep only unique entities
df['unq_GPE'] = df['GPE'].str.split(',').apply(pd.unique)
df['unq_FAC'] = df['FAC'].str.split(',').apply(pd.unique)
df['unq_LOC'] = df['LOC'].str.split(',').apply(pd.unique)
df['unq_ORG'] = df['ORG'].str.split(',').apply(pd.unique)
df[['GPE','unq_GPE']].loc[df['#GPE'] ==3]

df['unq_GPE'].dtype

# Data cleaning: Remove square brakets from location entities
df['unq_GPE'] =  df['unq_GPE'].astype(str).apply(lambda x: x.replace('[','').replace(']',''))
df['unq_FAC'] =  df['unq_FAC'].astype(str).apply(lambda x: x.replace('[','').replace(']',''))
df['unq_LOC'] =  df['unq_LOC'].astype(str).apply(lambda x: x.replace('[','').replace(']',''))
df['unq_ORG'] =  df['unq_ORG'].astype(str).apply(lambda x: x.replace('[','').replace(']',''))

df[['GPE','unq_GPE']].loc[df['#GPE'] ==3].head(10)

# Split with comma on unq entities
df['unq_GPE'] =  df['unq_GPE'].astype(str).replace('' '',',')
df[['GPE','unq_GPE']].loc[df['#GPE'] ==3].head(10)

df.to_csv('data//analysis_data//unique.csv')

# Final cleaned unique locations

df = pd.read_csv('data//analysis_data//unique_all_cleaned.csv')
df.head(3)

df =df.drop(columns =['Unnamed: 0', 'Unnamed: 0.1', 'Unnamed: 0.1.1','Unnamed: 0.1.1.1','Unnamed: 4'])
df.head(3)

# replace all NaN with empty string
df = df.replace(np.nan, '', regex=True)
df.head()

# Counting comma separated entities
df['#LOC2'] = df['unq_LOC'].str.count(',')
df['#GPE2'] = df['unq_GPE'].str.count(',')
df['#ORG2'] = df['unq_ORG'].str.count(',')
df['#FAC2'] = df['unq_FAC'].str.count(',')
df[['GPE','#GPE','unq_GPE','#GPE2']].head(10)

df.to_csv('data//analysis_data//unique_all_cleaned.csv')

df.columns

## Discard GPE > 1 comma

df[['FAC','unq_FAC','#FAC','#FAC2']].loc[df['#FAC'] >=2]

df.shape

df_final = df.loc[df['#GPE2'] <=1]
df_final.shape

df_final = df_final.loc[df['#LOC2'] <=1]
df_final.shape

df_final = df_final.loc[df['#FAC2'] <=1]
df_final.shape

df_final = df_final.loc[df['#ORG2'] <=1]
df_final.shape

df_final.columns

df_final = df_final.drop(columns = ['#LOC', '#GPE', '#ORG', '#FAC', 'GPE', 'FAC', 'ORG', 'LOC', '#LOC2', '#GPE2', '#ORG2', '#FAC2'])
df_final.columns

df_final = df_final.rename(columns = {'unq_GPE':'GPE','unq_FAC':'FAC', 'unq_LOC':'LOC', 'unq_ORG':'ORG' })
df_final.columns

df_final.head(10)

df_final.to_csv('geocoding//data//final_posts.csv')

# Preparing geocoding locations

df = pd.read_csv('geocoding//data//final_posts.csv')
df.shape

# Print number of unique entities
print("Number of unique values:")
print("\n GPE =", df['GPE'].nunique(), "\n ORG =", df['ORG'].nunique(), "\n LOC =", df['LOC'].nunique(), "\n FAC =", df['FAC'].nunique())

# replace all NaN with empty string
df = df.replace(np.nan, '', regex=True)
df.head()

df['LOC'].unique()

# Combining locational entities to get finer and more informed place names
# Locations are combined only when both columns are not null

df['FAC_GPE'] = np.where(((df['FAC'] != '') & (df['GPE'] != '')), df['FAC'].str.cat(df['GPE'], sep = ", "), '')
df['ORG_GPE'] = np.where(((df['ORG'] != '') & (df['GPE'] != '')), df['ORG'].str.cat(df['GPE'], sep = ", "), '')
df['LOC_GPE'] = np.where(((df['LOC'] != '') & (df['GPE'] != '')), df['LOC'].str.cat(df['GPE'], sep = ", "), '')

df['FAC_LOC'] = np.where(((df['FAC'] != '') & (df['LOC'] != '')), df['FAC'].str.cat(df['LOC'], sep = ", "), '')
df['ORG_LOC'] = np.where(((df['ORG'] != '') & (df['LOC'] != '')), df['ORG'].str.cat(df['LOC'], sep = ", "), '')

df['LOC_FAC'] = np.where(((df['LOC'] != '') & (df['FAC'] != '')), df['LOC'].str.cat(df['FAC'], sep = ", "), '')
df['FAC_ORG'] = np.where(((df['FAC'] != '') & (df['ORG'] != '')), df['FAC'].str.cat(df['ORG'], sep = ", "), '')

df['ORG_FAC'] = np.where(((df['ORG'] != '') & (df['FAC'] != '')), df['ORG'].str.cat(df['FAC'], sep = ", "), '')
df['LOC_ORG'] = np.where(((df['LOC'] != '') & (df['ORG'] != '')), df['LOC'].str.cat(df['ORG'], sep = ", "), '')
df.head(2)

df['FAC_LOC_GPE'] = np.where(((df['FAC_LOC'] != '') & (df['GPE'] != '')), df['FAC_LOC'].str.cat(df['GPE'], sep = ", "), '')
df['ORG_LOC_GPE'] = np.where(((df['ORG_LOC'] != '') & (df['GPE'] != '')), df['ORG_LOC'].str.cat(df['GPE'], sep = ", "), '')
df['LOC_FAC_GPE'] = np.where(((df['LOC_FAC'] != '') & (df['GPE'] != '')), df['LOC_FAC'].str.cat(df['GPE'], sep = ", "), '')
df['FAC_ORG_GPE'] = np.where(((df['FAC_ORG'] != '') & (df['GPE'] != '')), df['FAC_ORG'].str.cat(df['GPE'], sep = ", "), '')
df['ORG_FAC_GPE'] = np.where(((df['ORG_FAC'] != '') & (df['GPE'] != '')), df['ORG_FAC'].str.cat(df['GPE'], sep = ", "), '')
df['LOC_ORG_GPE'] = np.where(((df['LOC_ORG'] != '') & (df['GPE'] != '')), df['LOC_ORG'].str.cat(df['GPE'], sep = ", "), '')
df.head(2)

df.shape

df.to_csv('Geocoding//data//final_posts.csv')

## Extract Unique locations to Geocode

unique_GPE = df.groupby('GPE')['id'].unique()
outfilename = ('geocoding//unq_locations//GPE.csv')
unique_GPE.to_csv(outfilename)

unique_FAC = df.groupby('FAC')['id'].unique()
outfilename = ('geocoding//unq_locations//FAC.csv')
unique_FAC.to_csv(outfilename)

unique_LOC = df.groupby('LOC')['id'].unique()
outfilename = ('geocoding//unq_locations//LOC.csv')
unique_LOC.to_csv(outfilename)

unique_ORG = df.groupby('ORG')['id'].unique()
outfilename = ('geocoding//unq_locations//ORG.csv')
unique_ORG.to_csv(outfilename)

unique_FAC_GPE = df.groupby('FAC_GPE')['id'].unique()
outfilename = ('geocoding//unq_locations//FAC_GPE.csv')
unique_FAC_GPE.to_csv(outfilename)

unique_LOC_GPE = df.groupby('LOC_GPE')['id'].unique()
outfilename = ('geocoding//unq_locations//LOC_GPE.csv')
unique_LOC_GPE.to_csv(outfilename)

unique_ORG_GPE = df.groupby('ORG_GPE')['id'].unique()
outfilename = ('geocoding//unq_locations//ORG_GPE.csv')
unique_ORG_GPE.to_csv(outfilename)

unique_FAC_LOC_GPE = df.groupby('FAC_LOC_GPE')['id'].unique()
outfilename = ('geocoding//unq_locations//FAC_LOC_GPE.csv')
unique_FAC_LOC_GPE.to_csv(outfilename)

unique_FAC_ORG_GPE = df.groupby('FAC_ORG_GPE')['id'].unique()
outfilename = ('geocoding//unq_locations//FAC_ORG_GPE.csv')
unique_FAC_ORG_GPE.to_csv(outfilename)

unique_LOC_FAC_GPE = df.groupby('LOC_FAC_GPE')['id'].unique()
outfilename = ('geocoding//unq_locations//LOC_FAC_GPE.csv')
unique_LOC_FAC_GPE.to_csv(outfilename)

unique_ORG_LOC_GPE = df.groupby('ORG_LOC_GPE')['id'].unique()
outfilename = ('geocoding//unq_locations//ORG_LOC_GPE.csv')
unique_ORG_LOC_GPE.to_csv(outfilename)

unique_ORG_FAC_GPE = df.groupby('ORG_FAC_GPE')['id'].unique()
outfilename = ('geocoding//unq_locations//ORG_FAC_GPE.csv')
unique_ORG_FAC_GPE.to_csv(outfilename)

unique_LOC_ORG_GPE = df.groupby('LOC_ORG_GPE')['id'].unique()
outfilename = ('geocoding//unq_locations//LOC_ORG_GPE.csv')
unique_LOC_ORG_GPE.to_csv(outfilename)
