==> spaCy_OSM_GoogleMaps/8. Displacement.ipynb <==
## Join geocoded locations and compute Dispalcement 

# Loading required packages for geocoding with Nominatim

import geopandas as gpd 
import geopy
import matplotlib.pyplot as plt
from functools import partial 
from geopy import distance
from geopy.distance import geodesic
from tqdm import tqdm, tqdm_notebook # progress bar

#initiate 
tqdm.pandas()
import pandas as pd, numpy as np

#Defining files to be geocoded
unique_files = ['FAC.csv', 'GPE.csv','ORG.csv','LOC.csv', 'FAC_GPE.csv','LOC_GPE.csv', 'ORG_GPE.csv']

# Defining the address columns within corresponding files 
address_column_names = [elem.split('.')[0] for elem in unique_files]

results = 'Geocoding/displacement'
ressources = 'Geocoding/nom_coords'
for unique_file  in unique_files:
    ressource_path =  ressources +'/nom_geocoded_' + unique_file
    result_path =  ressources +'/dist_nom_' + unique_file
    df = pd.read_csv(ressource_path)
    df['distance_google'] = df.apply(lambda row: distance_calc (row),axis=1)
    df.to_csv(result_path)

df = pd.read_csv('Geocoding//geocoded_joined//nom_LOC_full.csv')
df.head(2)

# Distance function

def distance_calc (row):
    start = (row['geom_lat'], row['geom_long'])
    stop = (row['nlat'], row['nlong'])
    try:
        return geodesic(start, stop).km
    except ValueError:
        return np.nan
# try, except used to catch error when computing Null values on coordinates

df['distance_google'] = df.apply(lambda row: distance_calc (row),axis=1)
df.head(1)

df.to_csv('Geocoding//displacement//dist_nom_LOC_full.csv')
