==> spaCy_OSM_GoogleMaps/8. Merge coordinates.ipynb <==
# Join Google Maps locations

import pandas as pd 
import warnings 
warnings.filterwarnings('ignore')

df = pd.read_csv('Geocoding//data//final_posts.csv')
df1 = pd.read_csv('Geocoding//google_coords//google_LOC_FAC_GPE.csv')
print(df.shape, df1.shape)

df2 = pd.merge(df, df1, left_on='LOC_FAC_GPE', right_on = 'input_string')
df2 = df2[['id','date','source','TE','lang','en_text','LOC_FAC_GPE', 'formatted_address', 'geom_long', 'geom_lat','lat','long', 'number_of_results']]
df2.head()

df2.shape

df2.to_csv('Geocoding//geocoded_joined//google_LOC_FAC_GPE_full.csv')

# Join Nominatim

import pandas as pd 
import warnings 
warnings.filterwarnings('ignore')

df = pd.read_csv('Geocoding//data//final_posts.csv')
df1 = pd.read_csv('Geocoding//nom_coords//nom_LOC.csv')
print(df.shape, df1.shape)

df1 = df1.drop(0)

df2 = pd.merge(df, df1, on='LOC')
df2 = df2[['id_x','date','source','TE','lang','en_text','LOC','location', 'geom_long', 'geom_lat','nlat','nlong']]
df2.head()

df2.shape

df2.to_csv('Geocoding//geocoded_joined//nom_LOC_full.csv')
