==> spaCy_OSM_GoogleMaps/8.1 Date time conversion.ipynb <==
# Date conversions

pip install timefhuman

from timefhuman import timefhuman
import datetime, pandas as pd, numpy as np

df = pd.read_csv('data//analysis_data//keyword+temp_filtered_posts.csv')
df.columns

df = df.drop(columns = ['Unnamed: 0', 'Unnamed: 0.1', 'Unnamed: 0.1.1.1'])

df.TE = df.TE.replace(np.nan, '', regex=True)
df.head(1)

def addDate(x):
    splitted_date = x['date'].split('/')
    now = datetime.datetime(int(splitted_date[1][0:4]),int(splitted_date[1]), int(splitted_date[0]),0,0)
    try:
        return timefhuman(x['TE'].lower(),now=now)
    except: 
        pass

df['converted_date'] = df.apply(addDate, axis = 1)
df.head(10)
